+++ lib/src/evaluator.dart
@@ -6,7 +6,6 @@ import 'package:darwin/src/genetic_algorithm.dart';
 import 'package:darwin/src/phenotype.dart';
 import 'package:darwin/src/result.dart';
 
-// TODO: can be implemented as an Isolate
 abstract class PhenotypeEvaluator<P extends Phenotype<G, R>, G,
     R extends FitnessResult> {
   Object? userData;